move code into separate calculate_kspace.py file
